Added an enemy script that follows the user.
I have implemented the script and added different characteristics, where if the enemy picks up a fruit then the player will lose points too. I have added a lives UI text and added the logic to it when the user collides with the enemy, thus the lives counter goes down.