Added the gave over segment that shows the users score and allows them to replay the game.
I have implemented the game over event that occurs when the user has no more lives or has run out of time. I have designed it so when the game is over, the user can still see the background as the game over is opposed on top of a transparent panel.