Adding more core functionality such as level select page, fade animations, and game won screen.
I created a seperate class for a fade function as opposed to using the animator. Also for the level select page, instead of creating a script for each individual level to point the buttons to the level, I created a method called select that can be declared with a string as the input of the level to point to.